Better Combination and Amount management
Made change in bet multiplication for each icon appear in the reel and provide valid result for bet output

+++ Assets/Scripts/SlotSymbol.cs
@@ -11,13 +11,13 @@ public class SlotSymbol : ScriptableObject
 
     [Header("Value / Payout (tweak as needed)")]
     [Tooltip("Base value for this symbol. You can combine this with multipliers to compute final payout.")]
-    public int baseValue = 1;
+    public float baseValue = 1;
 
     [Tooltip("Multiplier applied for 2-of-a-kind (optional). Final payout = bet * baseValue * twoMultiplier")]
-    public int twoMultiplier = 1;
+    public float twoMultiplier = 1;
 
     [Tooltip("Multiplier applied for 3-of-a-kind (optional). Final payout = bet * baseValue * threeMultiplier")]
-    public int threeMultiplier = 5;
+    public float threeMultiplier = 5;
 
     [Header("Spin probability")]
     [Tooltip("Higher weight → symbol appears more often when randomly chosen")]
